Hole Position Tweaks + Background Audio
Tweaked the position of the holes a bit more and made the title more apparent.

Added a background music which can be toggled on and off. We cannot autoplay it with how browsers work these days. They require users to interact with the website to have any kind of audio be played